Borrar,crear,traer funcionando de producto

+++ BochaStoreProyecto.Maui/ViewModels/ProductoViewModel/DetailsProductoViewModel.cs
@@ -14,31 +14,31 @@ namespace BochaStoreProyecto.Maui.ViewModels.ProductoViewModel
     
     public class DetailsProductoViewModel
     {
-        private readonly APIService _APIService;
+        public readonly APIService _APIService;
         public Producto _productoTraido { get; set; }
         public DetailsProductoViewModel(APIService apiService,Producto productoTraido)
         {
-            _productoTraido = productoTraido;
             _APIService = apiService;
-
+            _productoTraido = productoTraido;
         }
 
-        public ICommand ClickBorrarProducto=>
-        new Command(async () =>
-        {
-            var idProductoBorrar = _productoTraido.idProducto;
-            await _APIService.DeleteProducto(_productoTraido.idProducto);
-            await App.Current.MainPage.Navigation.PopAsync();
+        public ICommand ClickBorrarProducto =>
+            new Command(async () =>
+            {
+                if (_APIService != null && _productoTraido != null)
+                {
+                    await _APIService.DeleteProducto(_productoTraido.idProducto);
+                    await App.Current.MainPage.Navigation.PopAsync();
+                }
+            });
 
-        });
+        /*        public ICommand EditarProducto=>
+                new Command(async () =>
+                {
+                    var idProductoBorrar = _productoTraido.idProducto;
+                    await _APIService.DeleteProducto(idProductoBorrar);
+                    await App.Current.MainPage.Navigation.PopAsync();
 
-        public ICommand EditarProducto=>
-        new Command(async () =>
-        {
-            var idProductoBorrar = _productoTraido.idProducto;
-            await _APIService.DeleteProducto(idProductoBorrar);
-            await App.Current.MainPage.Navigation.PopAsync();
-
-        });
+                });*/
     }
 }
